refactor(federation): move CompositionOptions to composition module
`CompositionOptions` configures the public `compose()` entry point, not
internal merger state, so it belongs in `composition/mod.rs` next to the
functions it parameterizes. Also removes the indirect `merger::merge`
re-export path.
